adding stop get path

+++ migrations/20180313093655_tours.js
@@ -2,8 +2,8 @@ exports.up = function(knex, Promise) {
   return knex.schema.createTable("tours", table => {
     table.increments('id').primary();
     table.text('type');
-    table.integer('firstLatitude');
-    table.integer('firstLongitude');
+    table.float('firstLatitude');
+    table.float('firstLongitude');
     table.text('name');
     table.text('description');
     table.text('duration');
